Create tokens.py
This code is for acessing the tokens in the twitterstream.py.

You have to register into the twitter developer console and register a new application and then generate the tokens. These tokens can be used to acess the new application that you are going to create. Here we are going to just stream the twitter data and then perform sentiment analysis.