Create twitterstream.py
Here tokens are imported. The execution starts form the main Function creating TwitterClient class object. This class will Call the tweepy API to first get all the Authenticates the application and then API class is called from tweepy library. 
The other classes: 
TwitterStreamer is used to stream the tweet. Though this class is not being used, but we can use this class to stream any perons timeline or we can stream the tweeter with certain words as we like. 
TwitterListener class is also not used directly, but we can do any operation on data like storing in a file or preprocessing the data.
TweetAnalyser class is first used to convert the tweets and various data of the tweet into a dataframe. This data frame is then returned and we call another method in this class analyse_sentiment to analyse the sentiment of each tweet. 
The sentiment is returned as -1 for negative tweets, 0 for normal and 1 for positive tweets. TextBlob is used to analyse the sentiment.